# intelligence-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Independent Care Intelligence \u2014 synthetic skeleton",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ready Readyz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents": {
      "get": {
        "summary": "Agents",
        "operationId": "agents_v1_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Agents V1 Agents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/runs": {
      "post": {
        "summary": "Run",
        "operationId": "run_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_v1_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentId": {
        "type": "string",
        "enum": [
          "resident_history",
          "handover_draft"
        ],
        "title": "AgentId"
      },
      "Claim": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "sources"
        ],
        "title": "Claim"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Period": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "Period"
      },
      "RunRequest": {
        "properties": {
          "agent_id": {
            "$ref": "#/components/schemas/AgentId"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "default": "Summarise the recorded care"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "resident_id",
          "period"
        ],
        "title": "RunRequest"
      },
      "RunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "agent_id": {
            "$ref": "#/components/schemas/AgentId"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "state": {
            "type": "string",
            "enum": [
              "completed",
              "draft"
            ],
            "title": "State"
          },
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          },
          "provider": {
            "type": "string",
            "const": "fake",
            "title": "Provider",
            "default": "fake"
          },
          "agent_version": {
            "type": "string",
            "title": "Agent Version",
            "default": "0.1.0"
          },
          "gateway_version": {
            "type": "string",
            "title": "Gateway Version",
            "default": "synthetic-structured-v1"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data_cutoff": {
            "type": "string",
            "format": "date-time",
            "title": "Data Cutoff"
          },
          "coverage_complete": {
            "type": "boolean",
            "title": "Coverage Complete"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "resident_id",
          "period",
          "state",
          "generated_at",
          "data_cutoff",
          "coverage_complete",
          "claims",
          "warnings"
        ],
        "title": "RunResult"
      },
      "SourceRef": {
        "properties": {
          "source_system": {
            "type": "string",
            "const": "synthetic",
            "title": "Source System",
            "default": "synthetic"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_type",
          "source_id",
          "version"
        ],
        "title": "SourceRef"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}